prints size of various types

+++ hello_world/6-size.c
@@ -10,7 +10,7 @@ int main(void)
 printf("Size of char: %zu bytes\n", sizeof(char));
 printf("Size of an int: %zu bytes\n", sizeof(int));
 printf("Size of a long int: %zu bytes\n", sizeof(int));
-printf("Size of a long long int: %zu bytes\n", sizeof(int));
+printf("Size of a long long int: %ld bytes\n", sizeof(long int));
 printf("Size of a float: %zu bytes\n", sizeof(float));
 return (0);
 }
